x86_64 experimental work

+++ kernel/include/syscall-handler.h
@@ -26,7 +26,7 @@ int handle_syscall_unlink(char * name);
 int handle_syscall_wait(int * status);
 int handle_syscall_write(int file, char *ptr, int len);
 int handle_syscall_gettimeofday(struct timeval * p, void * z);
-int handle_unknown_syscall(uint32_t syscall_id, uint32_t param1, uint32_t param2, uint32_t param3);
+int handle_unknown_syscall(long syscall_id, long param1, long param2, long param3);
 
 
 } // namespace
